Fix integer bitrate division
PiperOrigin-RevId: 237028164

+++ library/core/src/main/java/com/google/android/exoplayer2/trackselection/BufferSizeAdaptationBuilder.java
@@ -352,7 +352,8 @@ public final class BufferSizeAdaptationBuilder {
       // buffer = slope * ln(bitrate) + intercept,
       // with buffer(minBitrate) = minBuffer and buffer(maxBitrate) = maxBuffer - hysteresisBuffer.
       bitrateToBufferFunctionSlope =
-          (maxBufferUs - hysteresisBufferUs - minBufferUs) / Math.log(maxBitrate / minBitrate);
+          (maxBufferUs - hysteresisBufferUs - minBufferUs)
+              / Math.log((double) maxBitrate / minBitrate);
       bitrateToBufferFunctionIntercept =
           minBufferUs - bitrateToBufferFunctionSlope * Math.log(minBitrate);
     }
